+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-v0.3.0, 2013-01-09
+v0.4.0, 2014-02-09
+------------------
+
+  * Added support for nested/inheritable datasets
+  * Warn user if `ApplicationSeeds.config=` is called after `ApplicationSeeds.dataset=`
+
+
+v0.3.0, 2014-01-09
 ------------------
 
   * Added support for Capistrano 3.x